voice/requirements.txt
anthropic>=0.40.0
faster-whisper>=1.0.0
pvporcupine>=3.0.0
pyaudio>=0.2.14
elevenlabs>=1.0.0
python-dotenv>=1.0.0
